Fix bug that stopped the script if an addon from preset is deleted
Steam doesn't support adding hidden/removed/banned addons of the workshop into collections, so it happened previously that when a preset has such a hidden addon enabled, then the script for that preset would crash once the loop tries to add the hidden addon into the collection. Fixed by putting every attempt to toggle an addon into a try-catch block.